Add ability to cancel LapJV solver.
This adds the ability to cancel the LapJV solver by calling `cancel()`
on a new `Cancellation` token that's returned by the `cancellation()`
method on the solver. This is intended to be used with another thread
which sets up a timer to cancel the solver after a timeout, to protect
against long runtimes.

When cancelled, the solver returns an error and does not make any of its
intermediate results available.

The cancellation token is only checked at a few points throughout the
solver, to minimize the overhead of checking in tight loops. This should
still allow reasonably fine-grained timeouts, but this will need to be
profiled.